revision de  error del dia   09/04/2021  encontramos tipos de datos en  base  32  para la entrega por lo que prensentamos errores  de  convercion  asi que ajustamos  a 64int   iniciamos compilacion del instalador

+++ PUFT_PRUEBA_001/class/MAIN_GENERA_ENTREGAS.cs
@@ -199,7 +199,7 @@ namespace PUFT_PRUEBA_001
                         try
                         {
                             cmd.CommandType = CommandType.StoredProcedure;
-                            cmd.Parameters.Add(new MySqlParameter("NEW_ENTREGA", MySqlDbType.Int32)).Value = resultado_entrega.NUEVA_ENTREGA;
+                            cmd.Parameters.Add(new MySqlParameter("NEW_ENTREGA", MySqlDbType.Int64)).Value = resultado_entrega.NUEVA_ENTREGA;
                             cmd.Parameters.Add(new MySqlParameter("SAP_USUARIO", MySqlDbType.Int32)).Value = USU_LOGISTICA;
                             DataTable dt_table = new DataTable();
                             MySqlDataAdapter APSTER = new MySqlDataAdapter(cmd);
@@ -480,8 +480,8 @@ namespace PUFT_PRUEBA_001
                         try
                         {
                             cmd.CommandType = CommandType.StoredProcedure;
-                            cmd.Parameters.Add(new MySqlParameter("NEW_ENTREGA", MySqlDbType.Int32)).Value = resultado_entrega.NUEVA_ENTREGA;
-                            cmd.Parameters.Add(new MySqlParameter("NEW_FACTURA", MySqlDbType.Int32)).Value = FACTURA;
+                            cmd.Parameters.Add(new MySqlParameter("NEW_ENTREGA", MySqlDbType.Int64)).Value = resultado_entrega.NUEVA_ENTREGA;
+                            cmd.Parameters.Add(new MySqlParameter("NEW_FACTURA", MySqlDbType.Int64)).Value = FACTURA;
                             DataTable dt_table = new DataTable();
                             MySqlDataAdapter APSTER = new MySqlDataAdapter(cmd);
